+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Histórico de cambios
 >Todos los cambios notables de este proyecto se documentarán en este archivo.
 
+## [v0.0.2] - 2021-01-30
+### Añadidos
+* Nuevas funciones en ***StringFunctions.php***.
+* Nuevas funciones en ***FileFunctions.php***.
+* Nueva librería ***DateTimeFunctions.php***.
+
+---
 ## [v0.0.1] - 2021-01-08
 > Primera versión.